backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
boto3==1.34.0
pymongo==4.5.0
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
